+++ include/hagl/bitmap.h
@@ -54,13 +54,13 @@ typedef struct {
     uint16_t height;
     uint8_t depth;
     hagl_window_t clip;
-    void (*put_pixel)(void *self, int16_t x0, int16_t y0, color_t color);
-    color_t (*get_pixel)(void *self, int16_t x0, int16_t y0);
-    color_t (*color)(void *self, uint8_t r, uint8_t g, uint8_t b);
+    void (*put_pixel)(void *self, int16_t x0, int16_t y0, hagl_color_t color);
+    hagl_color_t (*get_pixel)(void *self, int16_t x0, int16_t y0);
+    hagl_color_t (*color)(void *self, uint8_t r, uint8_t g, uint8_t b);
     void (*blit)(void *self, int16_t x0, int16_t y0, void *src);
     void (*scale_blit)(void *self, int16_t x0, int16_t y0, uint16_t w, uint16_t h, void *src);
-    void (*hline)(void *self, int16_t x0, int16_t y0, uint16_t width, color_t color);
-    void (*vline)(void *self, int16_t x0, int16_t y0, uint16_t height, color_t color);
+    void (*hline)(void *self, int16_t x0, int16_t y0, uint16_t width, hagl_color_t color);
+    void (*vline)(void *self, int16_t x0, int16_t y0, uint16_t height, hagl_color_t color);
 
     uint16_t pitch;
     uint32_t size;