+++ src/hagl_rectangle.c
@@ -44,7 +44,7 @@ SPDX-License-Identifier: MIT
 #define MAX(a,b) (((a)>(b))?(a):(b))
 
 void
-hagl_draw_rectangle_xyxy(void const *_surface, int16_t x0, int16_t y0, int16_t x1, int16_t y1, color_t color)
+hagl_draw_rectangle_xyxy(void const *_surface, int16_t x0, int16_t y0, int16_t x1, int16_t y1, hagl_color_t color)
 {
     const hagl_surface_t *surface = _surface;
 
@@ -82,7 +82,7 @@ hagl_draw_rectangle_xyxy(void const *_surface, int16_t x0, int16_t y0, int16_t x
 }
 
 void
-hagl_fill_rectangle_xyxy(void const *_surface, int16_t x0, int16_t y0, int16_t x1, int16_t y1, color_t color)
+hagl_fill_rectangle_xyxy(void const *_surface, int16_t x0, int16_t y0, int16_t x1, int16_t y1, hagl_color_t color)
 {
     const hagl_surface_t *surface = _surface;
 
@@ -129,7 +129,7 @@ hagl_fill_rectangle_xyxy(void const *_surface, int16_t x0, int16_t y0, int16_t x
 }
 
 void
-hagl_draw_rounded_rectangle_xyxy(void const *_surface, int16_t x0, int16_t y0, int16_t x1, int16_t y1, int16_t r, color_t color)
+hagl_draw_rounded_rectangle_xyxy(void const *_surface, int16_t x0, int16_t y0, int16_t x1, int16_t y1, int16_t r, hagl_color_t color)
 {
     const hagl_surface_t *surface = _surface;
 
@@ -203,7 +203,7 @@ hagl_draw_rounded_rectangle_xyxy(void const *_surface, int16_t x0, int16_t y0, i
 };
 
 void
-hagl_fill_rounded_rectangle_xyxy(void const *_surface, int16_t x0, int16_t y0, int16_t x1, int16_t y1, int16_t r, color_t color)
+hagl_fill_rounded_rectangle_xyxy(void const *_surface, int16_t x0, int16_t y0, int16_t x1, int16_t y1, int16_t r, hagl_color_t color)
 {
     const hagl_surface_t *surface = _surface;