+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ## [0.7.0-dev](https://github.com/tuupola/hagl/compare/0.6.0...master) - unreleased
 
-### Added
+### Changed
+- Rename bitmap_size() to hagl_bitmap_size() ([#99](https://github.com/tuupola/hagl/pull/99)).
 
+### Added
 - New `hagl_bitmap_init()` function ([#98](https://github.com/tuupola/hagl/pull/98)).
 
 ### Removed
-
 - Legacy `bitmap_init()` function ([#98](https://github.com/tuupola/hagl/pull/98)).